+++ auth/mock/auth.go
@@ -34,7 +34,7 @@ func (m *MockAuthClient) EXPECT() *MockAuthClientMockRecorder {
 }
 
 // GenerateToken mocks base method.
-func (m *MockAuthClient) GenerateToken(id int) (string, error) {
+func (m *MockAuthClient) GenerateToken(id int64) (string, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GenerateToken", id)
 	ret0, _ := ret[0].(string)
@@ -49,10 +49,10 @@ func (mr *MockAuthClientMockRecorder) GenerateToken(id interface{}) *gomock.Call
 }
 
 // ValidateToken mocks base method.
-func (m *MockAuthClient) ValidateToken(signedToken string) (int, error) {
+func (m *MockAuthClient) ValidateToken(signedToken string) (int64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ValidateToken", signedToken)
-	ret0, _ := ret[0].(int)
+	ret0, _ := ret[0].(int64)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
